All prepared to Tutorial (included mission system in multiplayer); TODO end tutorial and start lvl1

+++ Assets/Scripts/System/Missions/Mission.cs
@@ -7,7 +7,7 @@ public abstract class Mission : ScriptableObject
 {
     public int idMission;
     public bool Done;
-    public float Points;
+    public int Points;
     public string Title;
     public string Description;
     public GameEvent Event;
@@ -18,7 +18,7 @@ public abstract class Mission<T> : ScriptableObject
 {
     public int idMission;
     public bool Done;
-    public float Points;
+    public int Points;
     public string Title;
     public string Description;
     public GameEvent<T> Event;
@@ -29,7 +29,7 @@ public abstract class Mission<T1, T2> : ScriptableObject
 {
     public int idMission;
     public bool Done;
-    public float Points;
+    public int Points;
     public string Title;
     public string Description;
     public GameEvent<T1,T2> Event; 
@@ -40,20 +40,10 @@ public abstract class Mission<T1, T2, T3> : ScriptableObject
 {
     public int idMission;
     public bool Done;
-    public float Points;
+    public int Points;
     public string Title;
     public string Description;
     public GameEvent<T1,T2,T3> Event;
     public abstract GameEventListener<T1, T2,T3> initValues(GameObject gameObject);
     public abstract void execute(GameEvent<T1,T2,T3> gameEvent, T1 parameter1, T2 parameter2, T3 parameter3);
 }
-//public abstract class Mission<T1, T2, T3, T4> : ScriptableObject
-//{
-//    public int idMission;
-//    public bool Done;
-//    public float Points;
-//    public string Title;
-//    public string Description;
-//    public abstract void initValues();
-//    public abstract void execute(GameEvent<T1, T2, T3,T4> gameEvent, T1 parameter1, T2 parameter2, T3 parameter3, T4 parameter4);
-//}
